test/vm_setup: add wget to list of installed packages
This might be a bit confusing, but it seems some of the
distributions using "minimal" install do not come with
wget installed, and it is used frequently in our provisioning
scripts. I stumbled upon this issue with Centos7, but am
adding wget for all distros, just to be sure.

Signed-off-by: Karol Latecki <karol.latecki@intel.com>
Change-Id: I9621d5ce92014623d76d70c9e22d9eefa9ea0203
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/1994
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>